# docs/swagger.yaml
basePath: /api/public/accounting/v1
definitions:
...
  dto.CreateJournalBatchRequest:
    properties:
      batch_no:
        example: BATCH-20260415-001
        maxLength: 50
        type: string
      batch_type:
        example: SALARY_POSTING
        maxLength: 50
        type: string
      book_id:
        example: 550e8400-e29b-41d4-a716-446655440010
        type: string
      branch_id:
        example: 550e8400-e29b-41d4-a716-446655440080
        type: string
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      narration:
        example: April payroll run
        maxLength: 2000
        type: string
      posting_date:
        example: "2026-04-15T00:00:00Z"
        type: string
      source_module:
        example: PAYROLL
        maxLength: 50
        type: string
    required:
    - batch_no
    - batch_type
    - book_id
    - company_id
    - posting_date
    type: object
  dto.CreateJournalRequest:
    properties:
      book_id:
        example: 550e8400-e29b-41d4-a716-446655440010
        type: string
      branch_id:
        example: 550e8400-e29b-41d4-a716-446655440080
        type: string
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      currency_code:
        example: INR
        type: string
      exchange_rate:
        example: 1
        minimum: 0
        type: number
      fiscal_year_id:
        example: 550e8400-e29b-41d4-a716-446655440020
        type: string
      idempotency_key:
        example: inv-2026-001-post
        maxLength: 200
        type: string
      journal_date:
        example: "2026-04-15T00:00:00Z"
        type: string
      journal_kind:
        enum:
        - MANUAL
        - OPENING
        - ADJUSTMENT
        - REVERSAL
        - SYSTEM
        - CLOSING
        example: MANUAL
        type: string
      lines:
        items:
          $ref: '#/definitions/dto.JournalLineRequest'
        minItems: 1
        type: array
      narration:
        example: Monthly accruals
        maxLength: 2000
        type: string
      period_id:
        example: 550e8400-e29b-41d4-a716-446655440030
        type: string
      posting_date:
        example: "2026-04-15T00:00:00Z"
        type: string
      reference_no:
        example: REF-7788
        maxLength: 80
        type: string
      source_document_id:
        example: 550e8400-e29b-41d4-a716-446655440100
        type: string
      source_document_type:
        example: INVOICE
        maxLength: 80
        type: string
      source_event_id:
        example: 550e8400-e29b-41d4-a716-446655440101
        type: string
      source_module:
        example: BILLING
        maxLength: 50
        type: string
    required:
    - book_id
    - company_id
    - currency_code
    - exchange_rate
    - fiscal_year_id
    - journal_date
    - journal_kind
    - lines
    - period_id
    - posting_date
    type: object
...
  dto.JournalLineRequest:
    properties:
      account_id:
        example: 550e8400-e29b-41d4-a716-446655440070
        type: string
      branch_id:
        example: 550e8400-e29b-41d4-a716-446655440080
        type: string
      cost_center_id:
        example: 550e8400-e29b-41d4-a716-446655440083
        type: string
      credit_amount_base:
        example: 0
        minimum: 0
        type: number
      credit_amount_txn:
        example: 0
        minimum: 0
        type: number
      debit_amount_base:
        example: 10000.5
        minimum: 0
        type: number
      debit_amount_txn:
        example: 10000.5
        minimum: 0
        type: number
      description:
        example: Rent expense
        maxLength: 500
        type: string
      employee_id:
        example: 550e8400-e29b-41d4-a716-446655440082
        type: string
      exchange_rate:
        example: 1
        minimum: 0
        type: number
      income_head_id:
        example: 550e8400-e29b-41d4-a716-446655440084
        type: string
      party_id:
        example: 550e8400-e29b-41d4-a716-446655440081
        type: string
      tax_code_id:
        example: 550e8400-e29b-41d4-a716-446655440085
        type: string
    required:
    - account_id
    type: object
  dto.PostJournalRequest:
    properties:
      journal_id:
        example: 550e8400-e29b-41d4-a716-446655440110
        type: string
      posting_date:
        example: "2026-04-16T00:00:00Z"
        type: string
    required:
    - journal_id
    type: object
  dto.ReverseJournalRequest:
    properties:
      journal_id:
        example: 550e8400-e29b-41d4-a716-446655440110
        type: string
      narration:
        example: Reversal of incorrect accrual
        maxLength: 2000
        type: string
      reversal_date:
        example: "2026-04-20T00:00:00Z"
        type: string
    required:
    - journal_id
    type: object
...
  dto.UpdateJournalBatchRequest:
    properties:
      batch_type:
        example: SALARY_POSTING
        maxLength: 50
        type: string
      narration:
        example: April payroll (corrected)
        maxLength: 2000
        type: string
      posting_date:
        example: "2026-04-16T00:00:00Z"
        type: string
      source_module:
        example: PAYROLL
        maxLength: 50
        type: string
      status:
        enum:
        - OPEN
        - POSTED
        - FAILED
        - CANCELLED
        example: POSTED
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        items:
          type: string
        type: array
      message:
        type: string
      status:
        type: boolean
      status_code:
        type: integer
    type: object
host: localhost:8086
info:
  contact:
    email: support@queuebuster.co
    name: API Support
    url: http://www.queuebuster.co/support
  description: Enterprise accounting module for QueueBuster ERP
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QB Accounting Service API
  version: 1.0.0
paths:
...
  /journal-batches:
    post:
      consumes:
      - application/json
      description: Create a new journal batch
      parameters:
      - description: Journal batch data
        in: body
        name: journal_batch
        required: true
        schema:
          $ref: '#/definitions/dto.CreateJournalBatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a journal batch
      tags:
      - journal-batches
  /journal-batches/{id}:
    delete:
      description: Delete a journal batch
      parameters:
      - description: Journal batch ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete a journal batch
      tags:
      - journal-batches
    get:
      description: Get a journal batch by ID
      parameters:
      - description: Journal batch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a journal batch
      tags:
      - journal-batches
    put:
      consumes:
      - application/json
      description: Update a journal batch
      parameters:
      - description: Journal batch ID
        in: path
        name: id
        required: true
        type: string
      - description: Journal batch fields
        in: body
        name: journal_batch
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateJournalBatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update a journal batch
      tags:
      - journal-batches
  /journal-batches/book/{book_id}:
    get:
      description: List journal batches for a book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List journal batches by book
      tags:
      - journal-batches
  /journals:
    post:
      consumes:
      - application/json
      description: Create a new journal (typically draft)
      parameters:
      - description: Journal data with lines
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/dto.CreateJournalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a journal
      tags:
      - journals
  /journals/{id}:
    delete:
      description: Delete a journal
      parameters:
      - description: Journal ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete a journal
      tags:
      - journals
    get:
      description: Get a journal by ID; set include_lines=true to load journal lines
      parameters:
      - description: Journal ID
        in: path
        name: id
        required: true
        type: string
      - description: Include journal lines
        in: query
        name: include_lines
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a journal
      tags:
      - journals
  /journals/book/{book_id}:
    get:
      description: List journals for an accounting book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List journals by book
      tags:
      - journals
  /journals/period/{period_id}:
    get:
      description: List journals for an accounting period
      parameters:
      - description: Accounting period ID
        in: path
        name: period_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List journals by period
      tags:
      - journals
  /journals/post:
    post:
      consumes:
      - application/json
      description: Post a draft journal to the ledger
      parameters:
      - description: Post journal request
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/dto.PostJournalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Post a journal
      tags:
      - journals
  /journals/reverse:
    post:
      consumes:
      - application/json
      description: Create a reversing journal entry for a posted journal
      parameters:
      - description: Reverse journal request
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/dto.ReverseJournalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Reverse a journal
      tags:
      - journals
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"